+++ animations/finalVideo.py
@@ -1,28 +1,21 @@
-# from moviepy.video.io.VideoFileClip import VideoFileClip
-# import moviepy.editor as mp
 from moviepy.editor import *
+from animations.BaseClip import BaseClip
 
 
-# IMAGEMAGICK_BINARY = r'C:\Program Files\ImageMagick-7.0.8-Q16\magick.exe'
-
-
-class finalVideo:
+class finalVideo(BaseClip):
 
     def __init__(self):
         self.clip = None
-        self.cliplist = []
         self.duration = 0
 
-    def start_clip(self, clips):
+    def _start_clip(self, clip):
         """
         Slices a video file between two time frames to create a gif
         :param clip: Name of video file
-        :param start_time: Tuple stating time in (hour, min, sec), (min, sec) or (sec)
-        :param end_time: Tuple stating time in (hour, min, sec), (min, sec) or (sec)
         """
-        self.clip = CompositeVideoClip(clips=clips)
+        self.clip = CompositeVideoClip(clips=clip)
 
-    def add_clip(self, clip):
+    def _add_clip(self, clip):
         """
         Replace self.clip with a new CompositeVideoClip that starts with itself and the new clip.
 
@@ -42,123 +35,17 @@ class finalVideo:
         :param clip: Moviepy clip object
         :return: None
         """
-        self.cliplist.append(clip)
-
-    def build_clip(self):
-        '''
-        Build a final video clip by concatenating existing clips.
-
-        Builds a final video clip by going through the video clips in the
-            self.cliplist and adding them to the end of the current self.clip
-            CompositeVideoFile.
-        :return: None
-        '''
-        for clip in self.cliplist:
-            if self.duration is 0:
-                self.start_clip([clip.clip])
-                self.duration += clip.duration
-            else:
-                self.add_clip(clip.clip.set_start(self.duration))
-                self.duration += clip.duration
-
-    def resize(self, new_size):
-        """
-        Uses moviepy.video.fx.all.resize module
-        :param new_size: Can be wither(width,height) in pixels or a float
-                         A scaling factor, like 0.5
-                         A function of time returning one of these.
-        """
-        self.clip = self.clip.resize(new_size)
-
-    def crop(self, aspectRatio = None, x1=None, y1=None, x2=None, y2=None,
-             width=None, height=None, x_center= None, y_center= None):
-        """
-        Uses moviepy.video.fx.crop module. From documentation:
-        Returns a new clip in which just a rectangular subregion of the
-        original clip is conserved. x1,y1 indicates the top left corner and
-        x2,y2 is the lower right corner of the croped region.
-        All coordinates are in pixels. Float numbers are accepted.
-        :param x1: top left corner x-axis
-        :param y1: top left corner y-axis
-        :param x2: bottom right corner x-axis
-        :param y2: bottom right corner y-axis
-        :param width: width of rectangle
-        :param height: height of rectangle
-        :param x_center: x-axis center
-        :param y_center: y-axis center
-        """
-
-        # If a preselected aspect ratio was selected.
-        if aspectRatio:
-            if not x_center:
-                x_center = self.clip.w/2
-            if not y_center:
-                y_center = self.clip.h/2
-
-            # Vertical/Phone ratio
-            if aspectRatio == "vertical" or aspectRatio == "9:16" or aspectRatio == "phone":
-                self.clip = self.clip.crop(width=self.clip.h*9/16, height=self.clip.h,
-                                           x_center=x_center, y_center=y_center)
-
-            # Square ratio
-            elif aspectRatio == "square" or aspectRatio == "1:1":
-                self.clip = self.clip.crop(width=self.clip.h, height=self.clip.h,
-                                           x_center=x_center, y_center=y_center)
-
-            # 4:3/Letterbox ratio
-            elif aspectRatio == "4:3" or aspectRatio == "1.33:1" or aspectRatio == "letterbox":
-                self.clip = self.clip.crop(width=self.clip.h*1.33, height=self.clip.h,
-                                           x_center=x_center, y_center=y_center)
-
-            # 16:9/Widescreen ratio
-            elif aspectRatio == "16:9" or aspectRatio == "widescreen" or aspectRatio == "1.77:1":
-                    self.clip = self.clip.crop(width=self.clip.w, height=self.clip.w/1.77,
-                                               x_center=x_center, y_center=y_center)
-
-            # 21:9/Cinemascope ratio
-            elif aspectRatio == "cinemascope" or aspectRatio == "21:9" or aspectRatio == "2.33:1":
-                self.clip = self.clip.crop(width=self.clip.w, height=self.clip.w/2.33,
-                                           x_center=x_center, y_center=y_center)
-
-            # 2.35:1/Anamorphic ratio
-            elif aspectRatio == "anamorphic" or aspectRatio == "2.35:1":
-                self.clip = self.clip.crop(width=self.clip.w, height=self.clip.w/2.35,
-                                           x_center=x_center, y_center=y_center)
-
-            # 2.39:1/DCI ratio
-            elif aspectRatio == "DCI" or aspectRatio == "2.39:1":
-                self.clip = self.clip.crop(width=self.clip.w, height=self.clip.w/2.39,
-                                           x_center=x_center, y_center=y_center)
-
-            # 2.9:1/Digital IMAX ratio
-            elif aspectRatio == "Digital IMAX" or aspectRatio == "2.9:1":
-                self.clip = self.clip.crop(width=self.clip.w, height=self.clip.w/2.9,
-                                           x_center=x_center, y_center=y_center)
-
-            # If an invalid aspect ratio was specified, raise an exception.
-            else:
-                raise AttributeError("Invalid Aspect Ratio specified: '" + str(aspectRatio) + "'")
-
-        # If no preset ratio was selected, use other crop parameters.
+        if self.duration is 0:
+            self._start_clip([clip.clip])
         else:
-            self.clip = self.clip.crop(x1=x1, y1=y1, x2=x2, y2=y2,
-                                       width=width, height=height,
-                                       x_center=x_center, y_center=y_center)
+            self._add_clip(clip.clip.set_start(self.duration))
 
-    def writeClip(self, output):
-        # self.clip = self.clip.fx(concatenate([self.clip, self.clip.fx(vfx.time_mirror)]))
-        # self.clip = self.clip.crossfadein(self.clip.duration / 2)
-        # self.clip = (CompositeVideoClip([self.clip,
-        #                                  self.clip.set_start(self.clip.duration / 2),
-        #                                  self.clip.set_start(self.clip.duration)])
-        #              .subclip(self.clip.duration / 2, 3 * self.clip.duration / 2))
+        self.duration += clip.duration
 
-        self.clip.write_videofile(output)
-        # self.clip.reader.close()
-        # self.clip.audio.reader.close_proc()
-
-
-    def add_text(self, text, font_size, color, font, interline, pos, duration):
-        text = TextClip(text, fontsize=font_size, color=color,
-                        font=font, interline=interline).set_pos(pos).set_duration(duration)
-        self.clip = CompositeVideoClip([self.clip, text])
+    def writeVideo(self, filename):
+        """
+        Write the video to a file.
+        :param filename: name and format of output file.
+        :return:
+        """
+        self.clip.write_videofile(filename)
